VIC-35: Make all data accessible only per tenant added test for user-service
