+++ src/main/resources/application.properties
@@ -184,5 +184,5 @@ twoFactorAuth.consultant.enabled=false
 
 multitenancy.enabled=false
 tenant.service.api.url=http://tenantservice.default:8080
-#spring.jpa.properties.hibernate.ejb.interceptor=de.caritas.cob.userservice.api.service.TenantHibernateInterceptor
-#hibernate.current_session_context_class=org.hibernate.context.internal.ThreadLocalSessionContext
+spring.jpa.properties.hibernate.ejb.interceptor=de.caritas.cob.userservice.api.service.TenantHibernateInterceptor
+hibernate.current_session_context_class=org.hibernate.context.internal.ThreadLocalSessionContext
